LINEDINB/linkedin project/requirements.txt
python-dotenv=1111=1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
pillow==10.2.0